separate event view component
